Use a regex heuristic for the raw-text OTA encryption scan
