+++ esphome_device_builder/helpers/device_yaml/__init__.py
@@ -72,6 +72,7 @@ from ._parsing import (
     retarget_fallback_ap_ssid,
     safe_stat_key,
     yaml_has_api_encryption,
+    yaml_has_ota_encryption,
     yaml_has_top_level_block,
 )
 from ._resolve import EsphomeConfigUnavailableError, run_esphome_config
@@ -129,5 +130,6 @@ __all__ = [
     "run_esphome_config",
     "safe_stat_key",
     "yaml_has_api_encryption",
+    "yaml_has_ota_encryption",
     "yaml_has_top_level_block",
 ]
